Various upgrades (plume version). (#220)
* Multiple RSP upgrades.

- Add extended inverse view matrix command.
- Implement clip ratio for viewports.

* Bias transformation decomposition to be have a more similar rotation when scales without uniform signs are involved.

* Fix inspector coordinates by using the FB size from the VI.

* Change G_EX_ASPECT so that auto is 0 (the old behavior)

* Add a size variable to the output of file_to_c

* Fix D3D12 Validation errors and add discard for framebuffer manager.

* Add cache for tile copies.

* Skip rows in large loads that end up overwriting the same portions of TMEM.

* Do not use native samplers when tile copies are involved.

* Add first iteration of vertex texcoord interpolation and look at interpolation.

* Add look at processor.

* Remove usage of API viewports from renderer.

* Rework workload and presentation to use VI information for aspect ratio. Add remove black borders option.

* Remove half pixel threshold at the edge from VI Renderer.

* Update plume.

* Update matrix group macros.

* Update plume.

* Add projection matrix group ID to the debugger and change matrix groups to display as hex (#221)

* Add projection matrix group ID to the debugger and change matrix groups to display as hex

* Add remaining group fields to matrix group display in debugger

* Remove werror

* Update plume with allow tearing feature check.

* Enable SDL_VULKAN_ENABLED for plume.

* Update to plume with different options.

* Revert to pre-plume behavior for swap chain wait on Vulkan.

* Add resolve region capability and lower D3D12GraphicsCommandList requirements.

* Fix discard logic for tile reinterpretation.

* Fix aspect ratio group data retrieval on original framerate.

* Add tile caching mechanism for reinterpretations.

* Skip hash checks if tiles are specified on interpolate mode and not automatic.

* Fix tile matching logic.

* Make Application actually use the discardFBs argument.

* Fix VI Framebuffer size detection for cases where the VI is hidden using scale instead.

* Add support for float matrix command.

* Add support for extended position vertex buffer. (#227)

* Move Fix Rect LR to be done during parsing instead of high res rendering.

* Do not clear hash versions when clearing the replacement map.

* Track extended origins in scissor and viewport using the stack, preventing them from leaking to other base states.

* Update rt64_rsp.cpp

* Fix fullscreen on Linux

* Add extended vertex segment for vertex velocity.

* Specify destination vector.

* Adapt to new swap chain behavior and remove present wait workaround for NVIDIA. (#228)

* Adapt to new swap chain behavior and remove present wait workaround for NVIDIA.

* Update rt64_application.cpp

* Update plume.

* Add functions for manually running commands

* Add texcoord wrap point command.

* Turn off Retina.

* Update plume with Mailbox revert.

* Fix Vulkan validation error.

* Consider case where SwapChain does not need a resize.

* Point to RT64 branch that doesn't support residency sets.

* Add fallback path for Windows API creation. Add placeholder for Intel Vulkan fallback.

* Delete stray argument from print.

* Fix stale reference issue.

* Add D3D12 workaround for Intel 6th Gen.

* Update fallback list with very old AMD driver.

* Update plume to fix device address bug on Vulkan.

---------

+++ src/shaders/RSPWorldCS.hlsl
@@ -12,8 +12,8 @@ struct RSPWorldCB {
 };
 
 [[vk::push_constant]] ConstantBuffer<RSPWorldCB> gConstants : register(b0);
-Buffer<int> srcPos : register(t1);
-Buffer<int> srcVel : register(t2);
+Buffer<float> srcPos : register(t1);
+Buffer<float> srcVel : register(t2);
 Buffer<int> srcNorm : register(t3);
 Buffer<uint> srcIndices : register(t4);
 StructuredBuffer<float4x4> worldMats : register(t5);
